requirement.txt
numpy
pytorch
pygame
cv2